Split routes into several subfiles
